Dummy Tact parser
This change sets up a Menhir-based parser for Tact to be able
to iterate on language design and run experiments.

Currently, it doesn't do much. Mostly handles parsing simplistic
structs (with typed fields) and empty interfaces. It's mostly done
so that we can start iterating over it. The error messages have not
been written yet.

Here's an example of a .tact file:

```
struct Test {
  test: Type
}

interface Query {
}
```

As you can see, not very useful yet.