Changed the way syntax & lang tests operate
There are problems with the current way we're testing:

1. Alcotest requires a test listing ceremony. Minor, but annoying.
2. Constructing AST/Term trees by hand takes time and error-prone
3. Tests fail as true/false and don't show what's wrong. A lot of time
   is spent figuring this out.
4. AST/Term trees are not very readable and take a lot of space.

So, I am switching tests to ppx_expect that allows to compare results
of textual output. To produce this output, I am using sexplib and
ppx_sexp_conv to produce AST/Term converters into sexp.

Why sexp? Mostly I didn't want to do a handcrafted formatter, especially
considering the need for pretty-printing. So I took what was there.

You may notice that now Parser and Land are parameterized over Syntax,
and Syntax is parameterized over Located. This is done for a relatively
bening goal of NOT including `'a located` into sexp as they pollute it
heavily to the point of making them nearly useless. `ppx_sexp_conv` does
not allow "record field redirection" so instead I decided to
parameterize Syntax over Located.T which may or may not wrap values with
a location. Long-term this may be a great thing as we can now
parameterize the parser/syntax/lang over other things as well.

As for how to use ppx_expect-based tests in a very efficient workflow,
please refer to the updated README. These tests now spit out a diff
(allowing to see where's the error) and `dune promote` can be used to
quickly accept the diff if it is deemed to be correct.