Tact function call inlining where feasible (proof of concept)
This relies on passing immediate terms to the functions and implying
that this means that these functions can't return any other value
because there are no other sources of non-determinism.

That is unlikely to hold true as development goes further. Some
functions that would be legitimately runtime only would be detected as
compile-time inlineables.

We'd either to have to come up with further methods of analysis to
qualify functions for inlining or make all sources of non-determinism
explicit and therefore easy to detect.