Function interpreter should respect resolved reference stripping
Otherwise the resulting code is a mix between stripped and non-stripped,
which is a bit messy.

It does this by trying to strip `env` and see if it did change. If it
didn't, this means it already has been stripped and this has to be
respected.

As a result of this `Lang.env_from_program` has to be followed by
`Lang.eval_env` to evaluate compile-time functions.