Fix compile-time function invocation to not receive non-immediate terms
I've noticed that if I have `println(i)` (`i` coming from function's
arguments) it'll actually print the `Reference i` term instead of the
given number.

So I rewrote the interpretation a little bit to ensure we evaluate
functions that we can at the right time with the right references
resolved.

As a side note, testing using builtin of `println` is not a great way to
ensure everything works correctly as the order of the execution there is
not defined. So I've added an ability to (optionally) pass environment
and therefore populate scope with any testing functions we'd need.