Introduce compile-time TypeOf() function
It takes an expression and returns the type of it. This way one can
define a type using another expression:

```
let a = 1;

fn test() -> TypeOf(a) {
  a
}
```

This will help with future needs of meta-programming.

Do note that that this is an early naive implementation but it does pass
some initial testing. Future evolution should include availability of
function's parameter names both to the return clause and the body of the
function so that the type can be deriven from them. Something like:

```
fn test(t: Any) -> TypeOf(t) {
  t
}
```

Note that this is not yet possible, but we're getting there.