Add `val` and `case` keywords to `struct` and `union` members
There are a couple of reasons for this addition:

* Clarity of the syntax ("why do we use `fn`, `impl` [TBD] but nothing
  for members?")
* Giving users terminology to use (`val` is not perfect but we couldn't
  find anything better; `field` is perhaps a better fit but doesn't roll
  of the keyboard as easily)

This change also removes shorthand `struct{Ident,Ident}` syntax in an
expectation that it can be replaced with a compile-time `tuple(...)`
function [1]

Overall, it is not strictly necessary to introduce this syntax from
parser's perspective (it was doing without it just fine) but perhaps it
introduces more regularity to the syntax.

[1] The function might not be able to assign symbolic names to fields
unless we extend the compiler to capture references for builtin
functions.