Implement a new execution model
Instead of a restricted top level, allow any compile-time code to be
executed. This will change recursive definitions are to be handled (not
done in this change as of yet) but it will allow simpler linear
execution of the code and the execution models will work the same in
functions and in the top-level.

@p0lunin found a great way to do type conversion using `fold` visitors,
kudos!

I've slightly redesigned the error reporting subsystem with future
extensibility in mind.