Re-implement current compiler passes into a single pass
This way we're consolidating processing in one step making it less
brittle.

Perhaps even more importantly, while doing so, we're paying extra
attention to resolving references at the right time.

The idea that I've tried to go for here is that we do want to expand
references proactively in compile-time where we can but leave references
in place for code generation inside of functions for proper value
binding.

Chances are, this is still imperfect and there are some bugs lurking
here and there but I think it's a progress.