Replace `Int257` built-in type with a parameterized `Int` type
The `Int(bits)` type creates a constuctor function that takes an
`Integer` and creates a newtype that holds the integer. Currently, it'll
chop off extra bits from the constructed integer if it doesn't fit the
number of bits required. This may raise an error later, we don't have
much of an error harness there at the moment.

```
let counter = Int(257)(100);
```

It is important to note that I don't believe that the return of the
constructor function is how we'll end up structuring this interface,
as it currently would prevent us from using it as a type in signatures
and explicit types (because `Int(257)` would _currently_ be a function)
but I am working towards figuring out this part to make every piece fit
together nicely.

The importance of this change lies in supporting serialization and
deserialization of data types.