Make `Int(bits)` a parameterized type, not a function
Current way we deal with `Int(bits)` is not great as it does not really
make it a type but rather some kind of constructor function:

```
let a = Int(257)(100);
```

This change makes `Int(bits)` return a struct that contains an
initialization method `new` that returns an instance of a structure
containing an integer underneath:

```
let a = Int(257).new(100);
```

I've had to implements portions of method calling infrastructure from
parser to the compiler pass. It's not complete but simply allowed me to
get to this crude version of this type. Gaps to be filled and rough
edges to be corrected.

Also, this change does not deal with anything else related to this type,
such as automatic dereferencing (need `impl`s for that, perhaps) to
enable casting to `Integer` and other operations.