Quick-and-dirty JavaScript build prototype
It's not final and it doesn't produce anything really useful for now but
spits out a JSON to be parsed (if anybody ever needs it, as it would be
more practical to spit out compiled code, funC, asm, binary).

This is mostly to work out the mechanics of having a Javascript build.

```js
JSON.parse(require('./_build/default/bin/tact_js.bc').Tact.parse("let a = 1"));
{
  stmts: [ [ 'Let', [Array] ] ],
  bindings: [
    [ 'a', [Array] ],
    [ 'Integer', [Array] ],
    [ 'Int', [Array] ],
    [ 'Bool', [Array] ],
    [ 'Type', [Array] ],
    [ 'Void', [Array] ]
  ]
}
```

One known problem is that in the current configuration `dune utop`
fails:

```
Unimplemented Javascript primitive caml_js_expr!
```