Git rid of tuple version of struct_id
It was originally done so that `Int()` can generate unique IDs for
different parameters but this is completely wrong as it depends on the
fact that `Int()` was taking an int.

Instead, let types like `Int()` use a memoization technique to re-use
struct IDs if called with the same arguments.