Externalize methods in respect to types
This allows methods to refer to the type it is included in:

```
struct T {
  fn new() -> Self { ... }
  fn method(self: Self) { ... }
}
```

With this change, `struct_` no longer has `struct_methods` and instead a
lookup table of methods is created in the `program`.

Due to the nature of some builtin functions and how they return types
that are dependent on arguments, I've introduced optional dependent
typing for built-in functions. It's not beautiful and is likely to be
revised in the future, especially when type inference will be in full
force. But for now it'll do, I guess.