Resolve `Self` in interface definitions to `Type SelfType` values
I am not 100% sure this is to stay, but at least for now it allows us to
denote this case without pinning it to a specific type.

In the future this may lead to an introducing of `Self` value of this
type if they will be deemed useful or even necessary. We'll see.