Introduce `Config` abstraction for configuring the language.
It has been pretty annoying maintaining the signature of the
`Syntax.Make` functor. One needed to update type definitions as well as
the interface of the auto-generator visitor class every time syntax is
changed or amended. This is burdensome and is not a great developer
experience.

This change allows us to carry the configuration instead of the product
(`Syntax`) and drop the need for synchronizing `Syntax.Make`'s signature.