Add `with_bindings` functions.
The key problem with mutability is that it is easy to forgot about what was mutated and where.
To fix this, I add `with_bindings` function that change state s1 to s2, call function f, and then change state back from s2 to s1.
This change improve readibility and possible remove future bugs related to this issue.