Generalize allowed methods and impls definitions.
It is possible situation when method can depend on the input argument of generic function, like in the following:
```
fn Foo(X: Type) {
 struct {
  fn id(self: Self, x: X) -> X { x }
 }
}
```
In this case we need to postpound function generation to times when it will be possible to know the type.
This commit implements this. In this commit was added fields program.methods_defs and program.impls_def.
These fields contains yet unknown methods and impls that must be resolved later when function `Foo` will called.
Current implementation is bad, so we will try to found another solution.