Initial prototype of TonVM interpreter
We pursue the following goals with this:

* Integration tests for testing actual behaviour of Tact code
  compiled through the upcoming TonVM codegen backend
* Instruction set ready for the TonVM codegen backend
* Establishing strong understanding of TonVM semantics
  expressed through a concise, unoptimized implementation

Current code dump is more of an early draft meant to establish some
rough implementation guidelines and can (will) evolve over time. It
covers only a small fraction of instructions at this time. The idea is
that we'll keep adding these as well as improving the interpreter over
time while we're preparing the other parts of Tact.