Base implementation of interface constraints
This commit implements possiblity to use interfaces as types as `Type`.
When interface placed in the type place, it say that "this value can accept
only types that implements this interface". For example, in the following example
`T` has type `Beep`, so `test(IT)` will accept only types `IT` that implements
`Beep` interface:
```
fn test(T: Beep) {
   fn(arg: T) {
      return arg.beep();
   }
}
```
This is requires special handling of interfaces when building method calls.
This is also requires new expr - IntfMethodCall which points that in that point
it is unknown which concrete function should be called, and this will be known
only when concrete type will be provided. Partial Evaluator will then check if the
interface from a IntfMethodCall is known. If it is known then it will be replaced
with MethodCall with concrete method got from concrete implementation.

Also fix bug with InvalidType of MkFunction

As a side artifact, compile-time evaluating was broken in tests/builtin.ml, but I
do not know why.