Split tonvm extended instruction sets from the core ones
These extended instruction sets typically contain aliases or shorthand
mnemonics for certain cases.

We do want to support convenient forms, especially when it comes to
using these in codegen (SWAP is easier to understand than XCHG0 1) but
not at the expense of a bloated VM interpreter.

This separation allows us to express aliases/shorthands for what they
are and interpret only the minimalist instruction set.