Preliminary attempt to handle tensor-like types for internal use
The main reason behind this type, as confusing it may be for the end
user (it has to do with the stack machine abstraction that we're not
leaking to Tact), we still need it to access TVM primitives (either
through FunC or TVM directly.

Current implementation is rather rough and supports only tensors of
arity of two but that can be relatively easily improved in the future as
necessary.

I've implemented `Divmod` primitive as a simple example of tensor use
(it returns two values).

I've started populating builtin structs (Tensor2 parameterized by types)
in the keyspace of negative type identifiers. This is rather a hack (it
would have been better to add them to a program, but I ran out of steam;
can do it later). We've used this approach before.