Handle implicit returns (without the `return` keyword)
We'll process cases of breaks that lead to implicit returns
to accomplish this.

Our parser replaces the last non-semicolon expression in a block
as a break. This is done so that it can be distinguished from a
statement. It is therefore called an implicit return and can return from
a block or a function.

This change handles the case where we need to actually generate an
explicit return from a function that is not done explicitly.