Fix issue with SelfType.
Currently, this is not a fix at all but just a way to "shut up" the compiler.
The issue is that when `X.call()` was occured where `X: SomeInterface` then the compiler place
method definition from the interface which contains `SelfType`. I have not find a way
to remove this SelfType with current execution model.