Basic REPL for Tact for quick experiments
When invoked without any arguments, `tact` will now work as a REPL.
Please note that printing facilities are still very limited and a lot of
values are not printed correctly, but we can address this as an ongoing
concern.

In terms of internal changes, this brings an optional `result` into
the program and allows `return` to be used in the top level.