Start introducing type parameter syntax
We're planning to facilitate type inference in cases like this:

```
fn pow2[n: Integer](x: Int[n]) -> Int[n] {
  x * x
}

pow2(x) // `n` is inferred from `x`'s type (`Int[...]`)
```

by signifying the fact that `n: Integer` is a type parameter using a
specialized bracket syntax. Internally this still works as a
compile-time functions, however, it'll carry contextual information that
will enable type inference.

Developing such an inference engine will take time, however, we want the
target language to be available to early adopters as early as possible
and therefore we want to provide such a syntax at this point in time.
This way the programs will be written in the syntax that will bring the
type inference down the road.