Remove program result and evaluation associated with it
This was originally done to enable REPL but it is wrong as all we need
is just interpretation of a given statement within a context.

In order to keep bindings properly updated, I've transitioned
constructor's bindings to a ref (as opposed to a mutable field) so that
the interpreter can update those. This will come in handy with upcoming
assignments (`var = value`) feature.

I've also added an ability to map `located` values