Initial support for attributes
At this point, the support is enabled for types (structs, unions,
interfaces, enums [limited to parser only as enums are not compiled
yet], struct fields [`val`] and `impl`s). I've considered adding
attributes to union cases but wasn't able to determine whether that is
necessary at this point.

An intersting byproduct of this change is that now struct and union
contents (items) can be arranged in any order. Before, it had to be in a
sequence:

For structs: fields, followed by functions, followed by impls
For unions: cases, followed by functions, followed by impls

However, the addition of attributes created shift/reduce conflicts in
the parser as it wasn't able to make a determination if the following
item is still of the same type or the follow-up type.

Therefore I rewrote that piece of the parser to handle them in any order
and that allowed it to handle attributes correctly.

Please note that this change has introduced a minor regression in some
of the error printing. I've put a FIXME there as I am unable to handle
it right now.