Initial bits of an alternative parser
Menhir parser generator proves to be somewhat difficult to deal with,
especially when backtracking of some sort is required. We also seem to
have a faily complex parser with multitude of expression kinds that
we're getting lost in.

Parser combinators is our light of hope to get this fixed. Also, easier
to deal with non-generated code.