Allow parameterizing syntax for struct T { ... }
```
struct T[A: Type] { }
```

The combinator I designed for this should be generic enough to support
other cases (such as let T[A] = ...) as well as similar cases like
unions.

The idea behind `parameterization` combinator is that it captures zero
or more[1] type parameters (`[T: Type, I: Integer]`, etc) and returns a
function another combinator that produces an `expr` can use to wrap that
expr in a type function.

[1] This allows using parameterization like this:

```
struct T[A: Type][B: A] { }
```

and it transforms into a function `T[A: Type]` that returns a function
that takes `[B: A]` that returns an empty struct.