Add `TypeCall` type.
This type is a type of returning value when function with brackets `[]` was called.
For example, with such declaration:
```
struct Test[X: Type] { ... }
```
The expression `Test[Integer]` will have type `TypeCalle {func: Test; args: [Integer]}`.
It is not changed nothing, but in the future it will allow adding type inference based
on this type.