Implement struct construction in the alt parser
It has affected the parsing of function return types. What is this?

```
fn x() -> Self {}
```

Is this a function with an empty body or does `x()` return `Self {}` as
a type? From the compiler's perspective, this is obviously nonsense.
However, the parser is not that smart.

So, for now, I made a quick hack disallowing the use of struct
construction in the return type as is. It's still possible to write it
as:

```
fn x() -> (Self{}) {}
```

(however nonsensical this is)

There is likely a nicer way around this, but it solves the problem for
now.