Eliminate Int[NN] and Uint[NN] type confusion
Oftentimes, the above types are interpreted by people with limited to
no Tact knowledge as arrays of Ints or Uints with a fixed size.

We eliminate this by introducing IntNN and UintNN types that internally
convert to those Int[NN] and Uint[NN] types, respectively.

Currently this is done in the parser for simplicity's sake.