Initial prototype for better error locations in ambiguous scenarios
Here's the problem. The way I set it up, all these `attempts` in the
parser are used for dual-purposed errors.

First class of errors is "this is not me" and this is where `attempt`
should be effected.

The second class of errors is "this is me, but wrong syntaxx" and this
is where we should simply fail without trying to backtrack.

Being somewhat tired towards the end of the day, I implemented this with
an exception. Feedback is welcome.